streamlit_app/requirements.txt
PyNiteFEA
shapely
matplotlib
papermodels